# docs/api/api.yaml
...
consumes:
- application/json
- application/xml
definitions:
  ApiError:
    properties:
      code:
        description: response code
        format: int64
        type: integer
        x-go-name: Code
      message:
        description: response message
        type: string
        x-go-name: Message
    type: object
    x-go-package: github.com/apisix/manager-api/internal/utils/consts
...
host: 127.0.0.1
info:
  description: Manager API directly operates ETCD and provides data management for APISIX, provides APIs for Front-end or other clients.
  license:
    name: Apache License 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
  title: API doc of Manager API.
paths:
  /apisix/admin/check_ssl_cert:
    post:
      description: verify SSL cert and key
      operationId: checkSSL
      parameters:
      - description: cert of SSL
        in: body
        name: cert
        required: true
        type: string
      - description: key of SSL
        in: body
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "0":
          description: SSL verify passed
          schema:
            $ref: '#/definitions/ApiError'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/ApiError'
  /apisix/admin/check_ssl_exists:
    post:
      description: check SSL exists or not by sni
      operationId: checkSSLExist
      parameters:
      - description: cert of SSL
        in: body
        name: cert
        required: true
        type: string
      - description: key of SSL
        in: body
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "0":
          description: SSL exists
          schema:
            $ref: '#/definitions/ApiError'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/ApiError'
...
  /apisix/admin/services:
    get:
      description: Return the service list according to the specified page number and page size, and can search by name
      operationId: getServiceList
      parameters:
      - description: page number
        in: query
        name: page
        type: integer
      - description: page size
        in: query
        name: page_size
        type: integer
      - description: name of service
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "0":
          description: list response
          schema:
            items:
              $ref: '#/definitions/service'
            type: array
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/ApiError'
  /apisix/admin/ssl:
    get:
      description: Return the SSL list according to the specified page number and page size, and can search by sni
      operationId: getSSLList
      parameters:
      - description: page number
        in: query
        name: page
        type: integer
      - description: page size
        in: query
        name: page_size
        type: integer
      - description: sni of SSL
        in: query
        name: sni
        type: string
      produces:
      - application/json
      responses:
        "0":
          description: list response
          schema:
            items:
              $ref: '#/definitions/ssl'
            type: array
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/ApiError'
  /apisix/admin/upstreams:
    get:
      description: Return the upstream list according to the specified page number and page size, and can search by name
      operationId: getSSLList
      parameters:
      - description: page number
        in: query
        name: page
        type: integer
      - description: page size
        in: query
        name: page_size
        type: integer
      - description: name of upstream
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "0":
          description: list response
          schema:
            items:
              $ref: '#/definitions/upstream'
            type: array
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/ApiError'
...
produces:
- application/json
- application/xml
schemes:
- http
- https
swagger: "2.0"